requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Security-Too>=5.7.1
Flask-WTF>=1.2.1
WTForms>=3.1.1
email-validator>=2.1.0
PyMySQL>=1.1.0
Plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
bcrypt>=4.1.1
